Move `ApplicationContext` and `SpringApplicationContext` to `utbot-framework`
